HTMell.Content.Markdown

Description

Loading HTMell content from markdown files

What is markdown content?

data MarkdownContent

A markdown content value consists of metadata and markdown data. The metadata is read as HTMell.Content.Metadata.Frontmatter, the markdown is just the source after the frontmatter data and will be translated to HTML via CMark when calling toHTML.

Constructors

MarkdownContent

Fields

meta :: Map String String

markdown :: String

Obtaining markdown content

readMarkdown :: String -> MarkdownContent

Build a MarkdownContent value from a source String.

loadMarkdownContent :: FilePath -> IO (Maybe MarkdownContent)

Read a MarkdownContent value from a file.